Separate app for PayButton
