+++ BTCPayServer/Services/Apps/AppType.cs
@@ -7,6 +7,7 @@ namespace BTCPayServer.Services.Apps
 {
     public enum AppType
     {
-        PointOfSale
+        PointOfSale,
+        PayButton
     }
 }
