+++ BTCPayServer/Views/Apps/ListApps.cshtml
@@ -1,4 +1,5 @@
-﻿@model ListAppsViewModel
+﻿@using BTCPayServer.Services.Apps
+@model ListAppsViewModel
 @{
     ViewData["Title"] = "Stores";
 }
@@ -48,10 +49,9 @@
                             <td>@app.AppName</td>
                             <td>@app.AppType</td>
                             <td style="text-align:right">
-                                @if (app.IsOwner)
+                                @if (app.IsOwner && app.AppType != AppType.PayButton.ToString())
                                 {
                                     <a asp-action="@app.UpdateAction" asp-controller="Apps" asp-route-appId="@app.Id">Settings</a><span> - </span>
-                                    <a asp-action="paybutton" asp-controller="Apps" asp-route-appId="@app.Id">Pay Button</a><span> - </span>
                                 }
                                 <a asp-action="@app.ViewAction" asp-controller="Apps" asp-route-appId="@app.Id">View</a><span> - </span>
                                 <a asp-action="DeleteApp" asp-route-appId="@app.Id">Remove</a>
