+++ BTCPayServer/Views/Apps/ViewPayButton.cshtml
